chore(runtime): shrink LayoutComponent per-instance memory (#13565) e776cb3cc5 * chore(runtime): convert booleans to flags and change some uint types
* chore(runtime): move render paths to struct

* chore(runtime): move m_proxy to a unique_pointer

* fix: overlapping layout and shapes correctly respect hit order

+++ include/rive/animation/state_machine_fire_action.hpp
@@ -5,7 +5,7 @@
 namespace rive
 {
 class StateMachineInstance;
-enum class StateMachineFireOccurance : int
+enum class StateMachineFireOccurance : uint8_t
 {
     atStart = 0,
     atEnd = 1
